Refactor reverse proxy configuration management
Removed DefaultConfiguration and introduced InitialConfiguration and InitialConfigurator classes. Updated Get.cs and Program.cs to utilize the new configuration structure. Static configuration in appsettings.Development.json has been eliminated in favor of programmatic configuration.
