+++ src/EnvironmentGateway/EnvironmentGateway.Api/GatewayConfiguration/InitialConfiguration.cs
@@ -1,0 +1,9 @@
+﻿using Yarp.ReverseProxy.Configuration;
+
+namespace EnvironmentGatewayApi.GatewayConfiguration;
+
+internal record InitialConfiguration
+{
+    public RouteConfig[] Routes { get; init; }
+    public ClusterConfig[] Clusters { get; init; }
+}
