+++ src/EnvironmentGateway/EnvironmentGateway.Api/appsettings.Development.json
@@ -4,24 +4,5 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
-  },
-  "ReverseProxy": {
-    "Routes": {
-      "route1": {
-        "ClusterId": "cluster1",
-        "Match": {
-          "Path": "{**catch-all}"
-        }
-      }
-    },
-    "Clusters": {
-      "cluster1": {
-        "Destinations": {
-          "destination1": {
-            "Address": "https://neuko-know-how.com/"
-          }
-        }
-      }
-    }
   }
 }
